Rename item with divider to item
