+++ apps/cookbook/src/app/showcase/list-showcase/list-showcase.component.html
@@ -12,19 +12,19 @@
 
 <div class="page-example">
   <div>
-    <h2>List with items using dividers (default):</h2>
-    <cookbook-code-viewer [html]="dividersExampleTemplate"></cookbook-code-viewer>
+    <h2>List with items:</h2>
+    <cookbook-code-viewer [html]="items"></cookbook-code-viewer>
   </div>
 
   <div>
-    <cookbook-iphone src="/examples/list/with-dividers" showExternalLink="true"></cookbook-iphone>
+    <cookbook-iphone src="/examples/list/with-items" showExternalLink="true"></cookbook-iphone>
   </div>
 </div>
 
 <div class="page-example">
   <div>
-    <h2>List with items with no dividers</h2>
-    <cookbook-code-viewer [html]="items"></cookbook-code-viewer>
+    <h2>List with items with no dividers:</h2>
+    <cookbook-code-viewer [html]="noDividersExampleTemplate"></cookbook-code-viewer>
   </div>
 
   <cookbook-iphone
@@ -35,7 +35,7 @@
 
 <div class="page-example">
   <div>
-    <h2>List with selectable items using dividers:</h2>
+    <h2>List with selectable items:</h2>
     <p>
       The selected item is highligted in bold by default. This behaviour can be suppressed by
       passing <code>[disableSelectionHighlight]="true"</code> to the list component.
